% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptation.R
\name{adaptation.detail}
\alias{adaptation.detail}
\title{Adaptation-detail function}
\usage{
adaptation.detail(adaptation.ids, geometry = TRUE)
}
\arguments{
\item{adaptation.ids}{is a list of adaptation ids}
}
\description{
This function queries the Adaptation Detail API:
}
\keyword{adaptation.detail}
