% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptation.R
\name{adaptation.projects}
\alias{adaptation.projects}
\title{Adaptation projects function}
\usage{
adaptation.projects(
  lookup.type,
  loc.type,
  lookup.arg,
  detail = TRUE,
  geometry = TRUE
)
}
\arguments{
\item{lookup.type}{is the lookup type}

\item{loc.type}{is the type of location used in lookup}

\item{lookup.arg}{is the location used for the lookup}
}
\description{
This function queries the Adaptation Summary API:
}
\keyword{adaptation.projects}
