% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmental.R
\name{environmental}
\alias{environmental}
\title{Environmental function}
\usage{
environmental(lookup.type, loc.type, lookup.arg, api, geometry = FALSE)
}
\arguments{
\item{lookup.type}{is the lookup type}

\item{loc.type}{is the type of location used in lookup}

\item{lookup.arg}{is the location used for the lookup}

\item{api}{is the specific environmental API to query
#' @param geometry is a boolean indicating whether geometry data should be returned}
}
\description{
This function queries the historic API:
}
\keyword{environmental}
