% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fema.R
\name{fema}
\alias{fema}
\title{FEMA function}
\usage{
fema(lookup.type, loc.type, lookup.arg)
}
\arguments{
\item{lookup.type}{is the lookup type}

\item{loc.type}{is the type of location used in lookup}

\item{lookup.arg}{is the location used for the lookup}
}
\description{
This function queries the FEMA API:
}
\keyword{fema}
