% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{historic}
\alias{historic}
\title{Historic function}
\usage{
historic()
}
\arguments{
\item{lookup.type}{is the lookup type}

\item{loc.type}{is the type of location used in lookup}

\item{lookup.arg}{is the location used for the lookup}

\item{detail}{is a boolean indicating whether detailed adaptaion project information should be returned}

\item{geometry}{is a boolean indicating whether geometry data should be returned}
}
\description{
This function queries the historic API:
}
\keyword{historic}
