% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_lookup.R
\name{location.lookup}
\alias{location.lookup}
\title{Location-lookup function}
\usage{
location.lookup(lookup, type, arg)
}
\arguments{
\item{lookup}{is the location lookup type}

\item{type}{is the location type for the lookup}

\item{arg}{is the location lookup argument}
}
\description{
This function constructs a location lookup:
}
\keyword{location.lookup}
