% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{probability}
\alias{probability}
\title{Probability function}
\usage{
probability(lookup.type, loc.type, lookup.arg, probability.api)
}
\arguments{
\item{lookup.type}{is the lookup type}

\item{loc.type}{is the type of location used in lookup}

\item{lookup.arg}{is the location used for the lookup}

\item{probability.api}{is the specific probability API (depth, chance, cumulative, count) to query}
}
\description{
This function queries the Adaptation Summary API:
}
\keyword{probability}
