% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_geometry.R
\name{process.geometry}
\alias{process.geometry}
\title{Process geometry function}
\usage{
process.geometry(parsed.data)
}
\arguments{
\item{parsed.data}{is the parsed response from the FSF API}
}
\description{
This function process geometry returns from the FSF APIs:
}
\keyword{process.geometry}
