% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_api_key.R
\name{set.api.key}
\alias{set.api.key}
\title{Set the First Street Foundation API key}
\usage{
set.api.key(key)
}
\arguments{
\item{key}{is the user's FSF key}
}
\description{
This function stores a user's First Street Foundation  (FSF) API key as the
package's environmental variable
}
