% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_version.R
\name{set.api.version}
\alias{set.api.version}
\title{Set the First Street Foundation API version}
\usage{
set.api.version(version)
}
\arguments{
\item{version}{is the user's FSF key}
}
\description{
This function stores the First Street Foundation  (FSF) API version
package's environmental variable
}
