% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{tile}
\alias{tile}
\title{Tile function}
\usage{
tile(product, year, return.period, event.id, z, x, y)
}
\arguments{
\item{product}{is the tileset product}

\item{year}{is the year of annual risk requested from the probability depth product}

\item{return.period}{is the return period requested from the probability depth product}

\item{event.id}{is the event id for the event tile requested from the historic API}

\item{z}{is the zoom level (min=1, max=18)}

\item{x}{is the X (latitude) coordinate (Mercator projection)}

\item{y}{is the Y (longitude) coordinate (Mercator projection)}
}
\description{
This function queries the historic API:
}
\keyword{tile}
