% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsf_query.R
\name{fsf.query}
\alias{fsf.query}
\title{FSF query function}
\usage{
fsf.query(api.cat, api, arg)
}
\arguments{
\item{api.cat}{is one of FSF's 7 API categories}

\item{api}{is one of FSF's 18 APIs}

\item{arg}{is a query argument}
}
\description{
This function constructs a fsf API request:
}
\keyword{fsf.query}
