% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_geometry.R
\name{process.geometry}
\alias{process.geometry}
\title{Process geometry function}
\usage{
process.geometry(geometry.data)
}
\arguments{
\item{geometry.data}{is the relevant geometry data from parsed API response}
}
\description{
This function process geometry returns from the FSF APIs:
}
\keyword{process.geometry}
